+++ packages/cli/__tests__/zosfiles/__unit__/delete/uss/Uss.handler.unit.test.ts
@@ -12,7 +12,6 @@
 import { Delete, IZosFilesResponse, IZosFilesOptions } from "@zowe/zos-files-for-zowe-sdk";
 import UssHandler from "../../../../../src/zosfiles/delete/uss/Uss.handler";
 import { ZosFilesBaseHandler } from "../../../../../src/zosfiles/ZosFilesBase.handler";
-
 describe("UssHandler", () => {
     const defaultReturn: IZosFilesResponse = {
         success        : true,