+++ packages/cli/__tests__/zosfiles/__unit__/hMigrate/HMigrate.definition.unit.test.ts
@@ -12,7 +12,6 @@
 import i18nEnglish from "../../../../src/zosfiles/-strings-/en";
 import { HMigrateDefinition } from "../../../../src/zosfiles/hMigrate/HMigrate.definition";
 import { ICommandDefinition } from "@zowe/imperative";
-
 describe("zos-files migrate group definition", () => {
     it ("should not have changed", () => {
         const definition: ICommandDefinition = require("../../../../src/zosfiles/hMigrate/HMigrate.definition").HMigrateDefinition;