+++ packages/cli/src/zosfiles/invoke/amsFile/AmsFile.handler.ts
@@ -12,7 +12,6 @@
 import { AbstractSession, IHandlerParameters, TextUtils } from "@zowe/imperative";
 import { IZosFilesResponse, Invoke, IZosFilesOptions } from "@zowe/zos-files-for-zowe-sdk";
 import { ZosFilesBaseHandler } from "../../ZosFilesBase.handler";
-
 /**
  * Handler to create a PDS data set
  * @export