+++ packages/workflows/__tests__/__system__/api/ListArchivedWorkflows.system.test.ts
@@ -91,7 +91,7 @@ describe("List archived workflows", () => {
             const response = await CreateWorkflow.createWorkflow(REAL_SESSION, wfName, definitionFile, system, owner);
             wfKey = response.workflowKey;
             // Archive workflow
-            await ArchiveWorkflow.archiveWorfklowByKey(REAL_SESSION, wfKey);
+            await ArchiveWorkflow.archiveWorkflowByKey(REAL_SESSION, wfKey);
         });
         afterEach(async () => {
             // deleting archived workflow