+++ packages/zosfiles/__tests__/__unit__/methods/create/Create.unit.test.ts
@@ -730,6 +730,73 @@ describe("Create data set", () => {
         );
     });
 
+    it("should be able to create a partitioned data set using the primary allocation and secondary allocation options", async () => {
+        const custOptions = {
+            dsorg: "PO",
+            alcunit: "CYL",
+            primary: 20,
+            secondary: 10,
+            dirblk: 5,
+            recfm: "FB",
+            blksize: 6160,
+            lrecl: 80
+        };
+        const response = await Create.dataSet(dummySession, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, dataSetName, custOptions);
+
+        expect(response.success).toBe(true);
+        expect(response.commandResponse).toContain("created successfully");
+        expect(mySpy).toHaveBeenCalledWith(
+            dummySession,
+            endpoint,
+            [],
+            JSON.stringify({
+                ...{
+                    alcunit: "CYL",
+                    dsorg: "PO",
+                    primary: 20,
+                    dirblk: 5,
+                    recfm: "FB",
+                    blksize: 6160,
+                    lrecl: 80,
+                    secondary: 10
+                }
+            })
+        );
+    });
+
+    it("should be able to create a partitioned data set using the primary allocation and default the secondary allocation", async () => {
+        const custOptions = {
+            dsorg: "PO",
+            alcunit: "CYL",
+            primary: 20,
+            dirblk: 5,
+            recfm: "FB",
+            blksize: 6160,
+            lrecl: 80
+        };
+        const response = await Create.dataSet(dummySession, CreateDataSetTypeEnum.DATA_SET_PARTITIONED, dataSetName, custOptions);
+
+        expect(response.success).toBe(true);
+        expect(response.commandResponse).toContain("created successfully");
+        expect(mySpy).toHaveBeenCalledWith(
+            dummySession,
+            endpoint,
+            [],
+            JSON.stringify({
+                ...{
+                    alcunit: "CYL",
+                    dsorg: "PO",
+                    primary: 20,
+                    dirblk: 5,
+                    recfm: "FB",
+                    blksize: 6160,
+                    lrecl: 80,
+                    secondary: 1
+                }
+            })
+        );
+    });
+
     describe("Expected failures", () => {
         it("should fail if the zOSMF REST client fails", async () => {
             const errorMsg = "Dummy error message";