+++ packages/zosfiles/src/index.ts
@@ -34,6 +34,7 @@ export * from "./methods/hRecall";
 export * from "./methods/invoke";
 export * from "./methods/list";
 export * from "./methods/mount";
+export * from "./methods/hRecall";
 export * from "./methods/rename";
 export * from "./methods/unmount";
 export * from "./methods/upload";
@@ -41,3 +42,7 @@ export * from "./methods/utilities";
 
 export * from "./utils/ZosFilesAttributes";
 export * from "./utils/ZosFilesUtils";
+export * from "./doc/IOptionsFullResponse";
+export * from "./doc/IRestClientResponse";
+export * from "./doc/IOptions";
+export * from "./doc/types/ZosmfRestClientProperties";