+++ packages/workflows/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the Zowe z/OS workflows SDK package will be documented in this file.
 
+## `6.25.0`
+
+- Enhancement: Deprecate misspelled version of `ArchiveWorkflow.archiveWorkflowByKey`
+- Bugfix: Remove "[object Object]" that appeared in some error messages
+
 ## `6.24.2`
 
 - Revert: Revert changes made in 6.24.1, problem was determined to be bundling pipeline